separated out halo scripts
